requirements.txt
flask
psutil
mysql-connector-python

